pyproject.toml
[tool.poetry]
name = "sample_agent"
version = "0.1.0"
description = "Starter"
authors = ["Luis Valdes <luis@valdes.com.br>"]
license = "MIT"

[project]
name = "sample_agent"
version = "0.0.1"
dependencies = [
    "crewai>=0.130.0",
    "ag-ui-crewai>=0.1.3",
    "ag-ui-protocol>=0.1.5",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.3"
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crewai = "0.130.0"
ag-ui-crewai = "0.1.3"
ag-ui-protocol = "0.1.5"
python-dotenv = "1.0.1"
uvicorn = "0.34.3"

[tool.poetry.scripts]
server = "sample_agent.main:main"